Tidy up code, comments and tests
